Moved all preprocessing in get_composite_source_model
