+++ openquake/commonlib/readinput.py
@@ -19,6 +19,7 @@ import os
 import re
 import ast
 import csv
+import time
 import copy
 import json
 import zlib
@@ -26,6 +27,7 @@ import pickle
 import shutil
 import zipfile
 import logging
+import operator
 import tempfile
 import functools
 import configparser
@@ -34,18 +36,21 @@ import collections
 import numpy
 import requests
 
-from openquake.baselib import hdf5, parallel
+from openquake.baselib import hdf5, parallel, performance
 from openquake.baselib.general import (
-    random_filter, countby, group_array, get_duplicates)
+    AccumDict, random_filter, countby, groupby, group_array, get_duplicates)
 from openquake.baselib.python3compat import zip
 from openquake.baselib.node import Node
 from openquake.hazardlib.const import StdDev
+from openquake.hazardlib.calc.filters import SourceFilter
 from openquake.hazardlib.calc.gmf import CorrelationButNoInterIntraStdDevs
 from openquake.hazardlib import (
     source, geo, site, imt, valid, sourceconverter, nrml, InvalidFile)
 from openquake.hazardlib.source import rupture
 from openquake.hazardlib.calc.stochastic import rupture_dt
 from openquake.hazardlib.probability_map import ProbabilityMap
+from openquake.hazardlib.source.point import grid_point_sources
+from openquake.hazardlib.sourceconverter import SourceGroup
 from openquake.risklib import asset, riskmodels
 from openquake.risklib.riskmodels import get_risk_functions
 from openquake.commonlib.oqvalidation import OqParam
@@ -708,25 +713,49 @@ def get_full_lt(oqparam):
     return full_lt
 
 
-def _get_cachedir(oq, full_lt, h5=None):
-    # read the composite source model from the cache
-    if not os.path.exists(oq.cachedir):
-        os.makedirs(oq.cachedir)
-    checksum = get_checksum32(oq, h5)
-    fname = os.path.join(oq.cachedir, 'csm_%s.pik' % checksum)
-    if os.path.exists(fname):
-        logging.info('Reading %s', fname)
-        with open(fname, 'rb') as f:
-            csm = pickle.load(f)
-            csm.full_lt = full_lt
-            return csm
-    csm = get_csm(oq, full_lt, h5)
-    if not csm.src_groups:  # everything was filtered away
-        return csm
-    logging.info('Saving %s', fname)
-    with open(fname, 'wb') as f:
-        pickle.dump(csm, f)
-    return csm
+def preclassical(srcs, params, monitor):
+    """
+    Prefilter and weight the sources. Also split if split_sources is true.
+    """
+    srcfilter = monitor.read('srcfilter')
+    # nrups, nsites, time, task_no
+    calc_times = AccumDict(accum=numpy.zeros(4, F32))
+    sources = []
+    for src in srcs:
+        t0 = time.time()
+        sources.extend(srcfilter.split_source(src, params['split_sources']))
+        dt = time.time() - t0
+        calc_times[src.id] += F32([src.num_ruptures, src.nsites, dt, 0])
+    for arr in calc_times.values():
+        arr[3] = monitor.task_no
+    dic = grid_point_sources(sources, params['ps_grid_spacing'])
+    dic['calc_times'] = calc_times
+    dic['before'] = len(sources)
+    dic['after'] = len(dic[sources[0].grp_id])
+    return dic
+
+
+def save_source_info(csm, h5):
+    data = {}  # src_id -> row
+    wkts = []
+    lens = []
+    for sg in csm.src_groups:
+        for src in sg:
+            lens.append(len(src.et_ids))
+            row = [src.source_id, src.grp_id, src.code,
+                   0, 0, 0, csm.full_lt.trti[src.tectonic_region_type], 0]
+            wkts.append(src._wkt)
+            data[src.id] = row
+    logging.info('There are %d groups and %d sources with len(et_ids)=%.2f',
+                 len(csm.src_groups), sum(len(sg) for sg in csm.src_groups),
+                 numpy.mean(lens))
+    csm.source_info = data  # src_id -> row
+    if h5:
+        attrs = dict(atomic=any(grp.atomic for grp in csm.src_groups))
+        # avoid hdf5 damned bug by creating source_info in advance
+        hdf5.create(h5, 'source_info', source_info_dt, attrs=attrs)
+        h5['source_wkt'] = numpy.array(wkts, hdf5.vstr)
+        h5['et_ids'] = csm.get_et_ids()
 
 
 def get_composite_source_model(oqparam, h5=None):
@@ -738,41 +767,75 @@ def get_composite_source_model(oqparam, h5=None):
     :param h5:
          an open hdf5.File where to store the source info
     """
-    logging.info('Reading the CompositeSourceModel')
+    # first of all, read the sites and the logic tree
+    sitecol = get_site_collection(oqparam, h5)
     full_lt = get_full_lt(oqparam)
+    srcfilter = SourceFilter(sitecol, oqparam.maximum_distance)
+    performance.Monitor.save(h5, 'srcfilter', srcfilter)
+
+    # then read the composite source model from the cache if possible
+    if oqparam.cachedir and not os.path.exists(oqparam.cachedir):
+        os.makedirs(oqparam.cachedir)
     if oqparam.cachedir and not oqparam.is_ucerf():
-        csm = _get_cachedir(oqparam, full_lt, h5)
-    else:
-        csm = get_csm(oqparam, full_lt,  h5)
-    et_ids = csm.get_et_ids()
+        checksum = get_checksum32(oqparam, h5)
+        fname = os.path.join(oqparam.cachedir, 'csm_%s.pik' % checksum)
+        if os.path.exists(fname):
+            logging.info('Reading %s', fname)
+            with open(fname, 'rb') as f:
+                csm = pickle.load(f)
+                csm.full_lt = full_lt
+                save_source_info(csm, h5)
+                return csm
+
+    # else read and process the composite source model
+    csm = get_csm(oqparam, full_lt,  h5)
     logging.info('%d effective smlt realization(s)', len(full_lt.sm_rlzs))
-    data = {}  # src_id -> row
-    mags_by_trt = csm.get_mags_by_trt()
-    wkts = []
-    lens = []
-    for sg in csm.src_groups:
-        for src in sg:
-            lens.append(len(src.et_ids))
-            row = [src.source_id, src.grp_id, src.code,
-                   0, 0, 0, full_lt.trti[src.tectonic_region_type], 0]
-            wkts.append(src._wkt)
-            data[src.id] = row
-    logging.info('There are %d groups and %d sources with len(et_ids)=%.2f',
-                 len(csm.src_groups), sum(len(sg) for sg in csm.src_groups),
-                 numpy.mean(lens))
-    if h5:
-        attrs = dict(atomic=any(grp.atomic for grp in csm.src_groups))
-        # avoid hdf5 damned bug by creating source_info in advance
-        hdf5.create(h5, 'source_info', source_info_dt, attrs=attrs)
-        h5['source_wkt'] = numpy.array(wkts, hdf5.vstr)
-        h5['et_ids'] = et_ids
-        for trt in mags_by_trt:
-            h5['source_mags/' + trt] = numpy.array(mags_by_trt[trt])
-        oqparam.maximum_distance.interp(mags_by_trt)
+    save_source_info(csm, h5)
+
+    # checks
     csm.gsim_lt.check_imts(oqparam.imtls)
-    csm.source_info = data  # src_id -> row
+
+    srcs = csm.get_sources()
+    if not srcs:
+        raise RuntimeError('All sources were discarded!?')
+
     if os.environ.get('OQ_CHECK_INPUT'):
         source.check_complex_faults(csm.get_sources())
+
+    logging.info('Checking the sources bounding box')
+    sids = srcfilter.within_bbox(srcs)
+    if len(sids) == 0:
+        raise RuntimeError('All sources were discarded!?')
+
+    # do nothing for atomic sources except counting the ruptures
+    for src in csm.get_sources(atomic=True):
+        src.num_ruptures = src.count_ruptures()
+        src.nsites = len(sitecol)
+
+    # run preclassical for non-atomic sources
+    sources_by_grp = groupby(
+        csm.get_sources(atomic=False), operator.attrgetter('grp_id'))
+    param = dict(ps_grid_spacing=oqparam.ps_grid_spacing,
+                 split_sources=oqparam.split_sources)
+    res = parallel.Starmap(
+        preclassical,
+        ((srcs, param) for srcs in sources_by_grp.values()), h5=h5,
+        distribute=None if len(sources_by_grp) > 1 else 'no').reduce()
+
+    if res and res['before'] != res['after']:
+        logging.info('Reduced the number of sources from {:_d} -> {:_d}'.
+                     format(res['before'], res['after']))
+
+    for grp_id, sources in sources_by_grp.items():
+        sg = SourceGroup(sources[0].tectonic_region_type)
+        sg.sources = res[grp_id]
+        csm.src_groups[grp_id] = sg
+
+    if oqparam.cachedir:
+        logging.info('Saving %s', fname)
+        with open(fname, 'wb') as f:
+            pickle.dump(csm, f)
+
     return csm
